Tests/LibWeb/Text/expected/DOM/cancelable-events.txt

event=test1 cancelable=false
event=test2 cancelable=true
event=test3 cancelable=false
event=test4 cancelable=true
